Refactor resolveIntellijVersions() as member method of IntellijVersions
