+++ buildSrc/src/main/kotlin/com/virtuslab/gitmachete/buildsrc/IntellijVersionHelper.kt
@@ -2,32 +2,8 @@ package com.virtuslab.gitmachete.buildsrc
 
 import java.io.File
 import java.util.*
-import kotlin.reflect.full.memberProperties
 
 object IntellijVersionHelper {
-  /**
-   * @param versionKey Either release number (like 2020.3) or key of intellijVersions (like
-   * eapOfLatestSupportedMajor)
-   * @returns Corresponding release numbers.
-   */
-  fun resolveIntelliJVersions(versionKey: String?, intellijVersions: IntellijVersions): List<String> {
-    if (versionKey == null) {
-      return emptyList()
-    }
-    val regex = "^[0-9].*$".toRegex()
-    if (regex.matches(versionKey)) {
-      return listOf(versionKey)
-    }
-
-    val propertiesList = listOfNotNull(
-      IntellijVersions::class.memberProperties
-        .single { it.name == versionKey }
-        .get(intellijVersions)
-    )
-
-    return propertiesList.flatMap { if (it is List<*>) it else listOf(it) }.map { it as String }
-  }
-
   fun getFromBuildNumber(buildNumber: String): String {
     return "20${buildNumber.substring(0, 2)}.${buildNumber.substring(2, 3)}"
   }
