+++ buildSrc/src/main/kotlin/com/virtuslab/gitmachete/buildsrc/UiTests.kt
@@ -20,7 +20,7 @@ fun Project.configureUiTests() {
 
   val uiTestTargets: List<String> =
     if (project.properties["against"] != null) {
-      IntellijVersionHelper.resolveIntelliJVersions(project.properties["against"] as String, intellijVersions)
+      intellijVersions.resolveIntelliJVersions(project.properties["against"] as String)
     } else {
       listOf(intellijVersions.buildTarget)
     }
